+++ packages/browser-wallet-api-helpers/src/wallet-api-types.ts
@@ -24,6 +24,7 @@ import type {
     ConfigureDelegationPayload,
     ContractName,
     EntrypointName,
+    TokenUpdatePayload,
 } from '@concordium/web-sdk';
 import type { RpcTransport } from '@protobuf-ts/runtime-rpc';
 import { LaxNumberEnumValue, LaxStringEnumValue } from './util';
@@ -204,6 +205,18 @@ interface MainWalletApi {
         accountAddress: AccountAddressSource,
         type: LaxNumberEnumValue<AccountTransactionType.TransferWithMemo>,
         payload: SimpleTransferWithMemoPayload
+    ): Promise<string>;
+    /**
+     * Sends a transaction to the Concordium Wallet and awaits the users action. Note that a header is not sent, and will be constructed by the wallet itself.
+     * Note that if the user rejects signing the transaction, this will throw an error.
+     * @param accountAddress the address of the account that should sign the transaction
+     * @param type the type of transaction that is to be signed and sent.
+     * @param payload the payload of the transaction to be signed and sent. Note that for smart contract transactions, the payload should not contain the parameters, those should instead be provided in the subsequent argument instead.
+     */
+    sendTransaction(
+        accountAddress: AccountAddressSource,
+        type: LaxNumberEnumValue<AccountTransactionType.TokenUpdate>,
+        payload: TokenUpdatePayload
     ): Promise<string>;
     /**
      * Sends a transaction to the Concordium Wallet and awaits the users action. Note that a header is not sent, and will be constructed by the wallet itself.