+++ packages/browser-wallet-api-helpers/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### Unreleased
+
+-   Updated handler `sendTransaction` with new payload type `TokenUpdatePayload`
+
 ## 3.0.1
 
 ### Fixed